+++ rollup.config.js
@@ -1,11 +1,7 @@
 import resolve from 'rollup-plugin-node-resolve';
-// import commonjs from 'rollup-plugin-commonjs';
 import babel from 'rollup-plugin-babel';
 import copy from 'rollup-plugin-copy'
-// import replace from 'rollup-plugin-replace';
-// import strip from 'rollup-plugin-strip';
-// import { uglify } from 'rollup-plugin-uglify';
-// import { sizeSnapshot } from 'rollup-plugin-size-snapshot';
+import { sizeSnapshot } from 'rollup-plugin-size-snapshot';
 import json from 'rollup-plugin-json';
 import pkg from './package.json';
 
@@ -41,10 +37,11 @@ export default [
       copy({
         targets: [
           'src/icons',
-          'src/composbale-icons.js',
+          'src/icons.js',
         ],
         outputFolder: 'dist',
-      })
+      }),
+      sizeSnapshot(),
     ],
   },
 ]